% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ablation.R
\docType{data}
\name{DIAG_KEYS}
\alias{DIAG_KEYS}
\title{Diagnostic keys matching DIAG_NAMES (length 15)}
\format{
An object of class \code{character} of length 15.
}
\usage{
DIAG_KEYS
}
\description{
Diagnostic keys matching DIAG_NAMES (length 15)
}
\keyword{internal}
