% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ablation.R
\docType{data}
\name{DIAG_NAMES}
\alias{DIAG_NAMES}
\title{Diagnostic display names (length 15)}
\format{
An object of class \code{character} of length 15.
}
\usage{
DIAG_NAMES
}
\description{
Diagnostic display names (length 15)
}
\keyword{internal}
