% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{.split_entry_counts}
\alias{.split_entry_counts}
\title{Split exiting entities into burst and normal entrants}
\usage{
.split_entry_counts(n_exits, burst_frac)
}
\arguments{
\item{n_exits}{Integer number of exiting entities.}

\item{burst_frac}{Probability an entrant is a burst entrant.}
}
\value{
Integer vector with \code{burst} and \code{normal} counts.
}
\description{
Split exiting entities into burst and normal entrants
}
\keyword{internal}
