% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankdiff-package.R
\name{.with_local_seed}
\alias{.with_local_seed}
\title{Evaluate an expression with a temporary RNG seed}
\usage{
.with_local_seed(seed, expr)
}
\arguments{
\item{seed}{Integer seed.}

\item{expr}{Expression to evaluate.}
}
\value{
The value of \code{expr}.
}
\description{
Evaluate an expression with a temporary RNG seed
}
\keyword{internal}
