% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{excess_kurtosis}
\alias{excess_kurtosis}
\title{Fisher excess kurtosis}
\usage{
excess_kurtosis(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Bias-corrected Fisher excess kurtosis.
}
\description{
Fisher excess kurtosis
}
\keyword{internal}
