% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{new_threshold_model}
\alias{new_threshold_model}
\title{Create a threshold model}
\usage{
new_threshold_model(
  threshold_by_period,
  max_missing_value_by_period,
  effectively_exact_above_threshold = TRUE
)
}
\arguments{
\item{threshold_by_period}{Numeric vector of per-period thresholds.}

\item{max_missing_value_by_period}{Numeric vector of max-missing values.}

\item{effectively_exact_above_threshold}{Logical.}
}
\value{
A list of class \code{rankdiff_threshold}.
}
\description{
Create a threshold model
}
\keyword{internal}
