% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{resolved_mc_reps}
\alias{resolved_mc_reps}
\title{Get resolved MC reps (respects dev_mode)}
\usage{
resolved_mc_reps(cfg)
}
\arguments{
\item{cfg}{A \code{rankdiff_config}.}
}
\value{
Integer number of MC reps.
}
\description{
Get resolved MC reps (respects dev_mode)
}
