% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\docType{data}
\name{SENSITIVITY_PARAMS}
\alias{SENSITIVITY_PARAMS}
\title{Default sensitivity parameters}
\format{
A list of length-2 character vectors.
}
\usage{
SENSITIVITY_PARAMS
}
\description{
A list of two-element character vectors, each giving
\code{c(display_name, attribute_name)} for a parameter to perturb.
}
\keyword{internal}
